♻️ refactor(webui): config/model 状态机下沉为领域 hook
候选 3：把 model.tsx 的 42 个 useState 状态机抽成领域 hook，主文件 2218→1338 行。

- 新增 model/hooks/useModelConfig：持有 models/api_providers/model_task_config 三份可编辑草稿 + loadConfig/saveConfig + 模型与提供商 CRUD + 搜索/分页/批量 + 表单校验 + 任务配置问题检查 + 连接测试 + 提供商删除级联（models↔providers↔taskConfig 深度耦合，合一个核心 hook 避免脆弱的跨 hook 回调）
- 新增 model/hooks/useEmbeddingWarning：用通用 usePendingOperation 收编 embedding 换模型的待定确认（previousEmbeddingModelsRef 一并下沉）
- model 数据是可编辑草稿（非只读服务端态），保留加载→草稿 + autosave 回写，不转 useQuery
- 子组件 ModelTable/ProviderList/TaskConfigCard/ProviderForm/Pagination 零改动；useModelAutoSave/Fetcher/Tour 复用
- 特征化测试 7/7 全绿（加载/embedding 警告/保存/连接测试/级联删除）
