+++ dashboard/src/routes/config/model/hooks/index.ts
@@ -5,3 +5,5 @@
 export { useModelAutoSave } from './useModelAutoSave'
 export { useModelTour } from './useModelTour'
 export { useModelFetcher, useAutoFetchModels } from './useModelFetcher'
+export { useModelConfig } from './useModelConfig'
+export { useEmbeddingWarning } from './useEmbeddingWarning'
